filing_processor_function/requirements.txt
SQLAlchemy
psycopg2-binary
google-generativeai